Implemented an absolute time variable (fElapsedTime)
This makes the moving objects on the screen always go at the same pace, no matter the framerate. This is useful not only in the actual application, but also helps with testing in the faster emulator on the computer.

+++ Projectile/Projectile.cpp
@@ -11,10 +11,10 @@ Projectile::~Projectile() {
   //??
 }
 
-short Projectile::getX() {
+float Projectile::getX() {
   return x;
 }
 
-short Projectile::getY() {
+float Projectile::getY() {
   return y;
 }
